requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.17.0
astroid==3.3.8
attrs==24.3.0
colorama==0.4.6
concurrent-log-handler==0.9.25
coverage==7.7.0
dill==0.3.9
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.2.4
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.1.0
numpy==2.2.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.0.0
propcache==0.2.1
psycopg==3.2.3
pylint==3.3.3
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
pywin32==308; platform_system == "Windows"
SQLAlchemy==2.0.44
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3
